route housekeeping
- moved main server components into /foundation and fixed all references
- exported authRoutes and recipeRoutes
- removed bcrypt unused files
- unit tests now click save on edited recipe

NEXT UP:
Save recipe route
