--- scripts/!quickStartWin/3.2-startServer.bat
@@ -1,4 +1,0 @@
-@echo off
-ECHO Starting CodefortheFences server
-cd c:/node/codeforthefences/server/
-npm start
