+++ server/package.json
@@ -18,7 +18,6 @@
   },
   "homepage": "https://github.com/Devin-M-D/cookbook#readme",
   "dependencies": {
-    "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "nodemailer": "^6.4.6",
